requirements.txt

Django==3.0
django-cors-headers==3.5.0
djangorestframework==3.12.1
